requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
plotly==5.22.0
google-generativeai==0.7.2
python-dotenv==1.0.1
matplotlib==3.9.0
scipy==1.13.1
Pillow==10.3.0
fpdf2==2.7.9
requests==2.32.3
kaggle==1.6.14
